+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add is_hummingbird flag for new SLA policy (OSIDB-4884)
 
 ## [5.10.1] - 2026-05-05
 ### Fixed